+++ src/RcppExports.cpp
@@ -30,8 +30,8 @@ BEGIN_RCPP
 END_RCPP
 }
 // udp_tokenise_tag_parse
-Rcpp::List udp_tokenise_tag_parse(SEXP udmodel, Rcpp::StringVector x, Rcpp::StringVector docid, std::string annotation_tokenizer, std::string annotation_tagger, std::string annotation_parser);
-RcppExport SEXP _udpipe_udp_tokenise_tag_parse(SEXP udmodelSEXP, SEXP xSEXP, SEXP docidSEXP, SEXP annotation_tokenizerSEXP, SEXP annotation_taggerSEXP, SEXP annotation_parserSEXP) {
+Rcpp::List udp_tokenise_tag_parse(SEXP udmodel, Rcpp::StringVector x, Rcpp::StringVector docid, std::string annotation_tokenizer, std::string annotation_tagger, std::string annotation_parser, int log_every, Rcpp::Function current_time);
+RcppExport SEXP _udpipe_udp_tokenise_tag_parse(SEXP udmodelSEXP, SEXP xSEXP, SEXP docidSEXP, SEXP annotation_tokenizerSEXP, SEXP annotation_taggerSEXP, SEXP annotation_parserSEXP, SEXP log_everySEXP, SEXP current_timeSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -41,7 +41,9 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< std::string >::type annotation_tokenizer(annotation_tokenizerSEXP);
     Rcpp::traits::input_parameter< std::string >::type annotation_tagger(annotation_taggerSEXP);
     Rcpp::traits::input_parameter< std::string >::type annotation_parser(annotation_parserSEXP);
-    rcpp_result_gen = Rcpp::wrap(udp_tokenise_tag_parse(udmodel, x, docid, annotation_tokenizer, annotation_tagger, annotation_parser));
+    Rcpp::traits::input_parameter< int >::type log_every(log_everySEXP);
+    Rcpp::traits::input_parameter< Rcpp::Function >::type current_time(current_timeSEXP);
+    rcpp_result_gen = Rcpp::wrap(udp_tokenise_tag_parse(udmodel, x, docid, annotation_tokenizer, annotation_tagger, annotation_parser, log_every, current_time));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -92,7 +94,7 @@ END_RCPP
 static const R_CallMethodDef CallEntries[] = {
     {"_udpipe_phrases_regex_locate", (DL_FUNC) &_udpipe_phrases_regex_locate, 3},
     {"_udpipe_udp_load_model", (DL_FUNC) &_udpipe_udp_load_model, 1},
-    {"_udpipe_udp_tokenise_tag_parse", (DL_FUNC) &_udpipe_udp_tokenise_tag_parse, 6},
+    {"_udpipe_udp_tokenise_tag_parse", (DL_FUNC) &_udpipe_udp_tokenise_tag_parse, 8},
     {"_udpipe_na_locf", (DL_FUNC) &_udpipe_na_locf, 1},
     {"_udpipe_udp_train", (DL_FUNC) &_udpipe_udp_train, 6},
     {"_udpipe_udp_evaluate", (DL_FUNC) &_udpipe_udp_evaluate, 6},